requirements.txt
colorama
pyyaml
plyer
dbus-python